Break up home into home+charts+why
