+++ pages/why.py
@@ -1,0 +1,52 @@
+import air
+from air_markdown import TailwindTypographyMarkdown as Markdown
+
+
+def render(request: air.Request):
+    return air.Children(air.Title("Air: The New FastAPI-Powered Python Web Framework (2025)"),
+        Markdown("""
+# Why use Air?
+
+- **Powered by FastAPI** - Designed to work with FastAPI so you can server your API and web pages from one app
+- **Fast to code** - Tons of intuitive shortcuts and optimizations designed to expedite coding HTML with FastAPI
+- **Air Tags** - Easy to write and performant HTML content generation using Python classes to render HTML
+- **Jinja Friendly** - No need to write `response_class=HtmlResponse` and `templates.TemplateResponse` for every HTML view
+- **Mix Jinja and Air Tags** - Jinja and Air Tags both are first class citizens. Use either or both in the same view!
+- **HTMX friendly** - We love HTMX and provide utilities to use it with Air
+- **HTML form validation powered by pydantic** - We love using pydantic to validate incoming data. Air Forms provide two ways to use pydantic with HTML forms (dependency injection or from within views)
+- **Easy to learn yet well documented** - Hopefully Air is so intuitive and well-typed you'll barely need to use the documentation. In case you do need to look something up we're taking our experience writing technical books and using it to make documentation worth boasting about
+
+**Documentation**: <a href="https://feldroy.github.io/air/" target="_blank">https://feldroy.github.io/air/</a>
+
+**Source Code**: <a href="https://github.com/feldroy/air" target="_blank">https://github.com/feldroy/air</a>
+
+
+## Installation
+
+Install using `pip install -U air` or `conda install air -c conda-forge`.
+
+For `uv` users, just create a virtualenv and install the air package, like:
+
+```sh
+uv venv
+source .venv/bin/activate
+uv add air
+uv add fastapi[standard]
+```
+
+## A Simple Example
+
+Create a `main.py` with:
+
+```python
+import air
+
+app = air.Air()
+
+
+@app.get("/")
+async def index():
+    return air.Html(air.H1("Hello, world!", style="color: blue;"))
+```
+""")
+)
