+++ templates/page.html
@@ -20,9 +20,11 @@
                 </div>
                 <div class="hidden md:flex items-center space-x-4">
                     <a href="/" class="px-3 py-2 rounded-md text-sm font-medium hover:bg-gray-700">Home</a>
+                    <a href="/why" class="px-3 py-2 rounded-md text-sm font-medium hover:bg-gray-700">Why?</a>
                     <a href="/about" class="px-3 py-2 rounded-md text-sm font-medium hover:bg-gray-700">About</a>
                     <a href="/concepts" class="px-3 py-2 rounded-md text-sm font-medium hover:bg-gray-700">Concepts</a>
                     <a href="/reference" class="px-3 py-2 rounded-md text-sm font-medium hover:bg-gray-700">Reference</a>
+                    <a href="/charts" class="px-3 py-2 rounded-md text-sm font-medium hover:bg-gray-700">Charts</a>
                 </div>
                 <div class="md:hidden flex items-center">
                     <label for="menu-toggle" class="cursor-pointer">
